Added interfaces and abstract classes, cleaned up classes
