+++ src/fees/FeeInt.java
@@ -1,0 +1,7 @@
+package fees;
+
+public interface FeeInt {
+
+    double getValue();
+    double calculateFee();
+}
