--- src/lists/MarketOrderListInt.java
@@ -1,9 +1,0 @@
-package lists;
-
-public interface MarketOrderListInt {
-
-    void setFirstMarketOrder();
-    void setMarketOrder(int i);
-    void buildOrderList();
-
-}
