+++ src/support/FirstMarketOrderCalculations.java
@@ -1,0 +1,8 @@
+package support;
+
+public interface FirstMarketOrderCalculations {
+
+    double calculateFirstMarketOrderVolume();
+
+    double calculateFirstMarketOrderPrice();
+}
