+++ src/support/NextMarketOrderCalculations.java
@@ -1,0 +1,9 @@
+package support;
+
+public interface NextMarketOrderCalculations {
+
+    double calculateNextMarketOrderVolume(double previousVolume);
+
+    double calculateNextMarketOrderPrice(double previousPrice);
+
+}
