+++ src/support/TradeDataInt.java
@@ -1,0 +1,14 @@
+package support;
+
+public interface TradeDataInt {
+
+    TradeType getTradeType();
+
+    double getTradeVolume();
+
+    int getNumberOfMarketOrders();
+
+    double getMinMarketOrderPrice();
+
+    double getMaxMarketOrderPrice();
+}
